requirements.txt
streamlit
langchain
langchain-groq
langchain-huggingface
langchain-community
chromadb
sentence-transformers  # Required for local embeddings fallback
pillow